Moved API and statistics routes to own files
Even more housekeeping
